requirement.txt
streamlit
pandas
matplotlib.pyplot
sqlite3
datetime